fix(api): bound failoverPolicy durations at admission; ConfirmationSamples *int32
failoverPolicy.minTerrestrialDwell and switchbackDelay were typed only `Format=duration`, with no
bound on the value. metav1.Duration parses negatives, and a negative minTerrestrialDwell SILENTLY
DISABLES the dwell — the runtime guard `now.Sub(LastSwitchback) < dwell` is never true for a negative
dwell, so a fat-fingered "-90s" turns off the anti-flap min-dwell with no error. A negative
switchbackDelay is equally meaningless, and an absurd value (e.g. "1000h") would pin terrestrial for
weeks. The doc advertised "30s-120s is a sane range" but nothing enforced even non-negativity.

Add a CEL x-kubernetes-validations rule to both duration fields bounding them to [0s, 24h] — 0 stays
valid (it is the documented "disabled" value), negatives are rejected, and 24h is a fat-finger ceiling
far above any LEO-pass-relative value. This matches how #231 tightened the failoverPolicy trigger
admission (bound the value at the API, not just at runtime). Regenerated into all four CRD copies
(config/crd, dist/chart, bundle/manifests, nephio).

Also change ConfirmationSamples from *int to *int32 (#203-Low, deferred by #220): the value is bounded
1-10 so it never overflowed either width, but *int32 is the Kubernetes API convention and is explicit
and cross-arch-deterministic. Ripples: the controller derefs with an int() cast; deepcopy regenerated;
the CRD schema gains format: int32.

Test: TestControllers "failoverPolicy duration admission (CEL)" runs against the envtest API server
with the REAL regenerated CRD — a negative or >24h minTerrestrialDwell/switchbackDelay is rejected as
Invalid with the field named, while 0s/normal/24h are admitted. This is the first duration CEL rule in
the API, so the envtest proves the apiserver actually enforces it (not just that controller-gen
accepted the marker). fail-on-empty confirms the specs ran.

+++ internal/controller/ntnslice_controller_test.go
@@ -419,7 +419,7 @@ var _ = Describe("NTNSlice Controller", func() {
 		AfterEach(func() { deleteSlice() })
 
 		It("holds terrestrial until N consecutive degraded samples confirm", func() {
-			confirmN := 2
+			confirmN := int32(2)
 			spec := baseSpec()
 			spec.FailoverPolicy.ConfirmationSamples = &confirmN
 			ntnSlice := &ntnv1alpha1.NTNSlice{
